refactor: moved scim middlewares to new file
